:recycle: Implement facade for ProcessInstance state handling
