+++ src/runtime/facades/index.ts
@@ -1,4 +1,5 @@
 export * from './flow_node_persistence_facade';
+export * from './process_instance_state_handling_facade';
 export * from './process_model_facade';
 export * from './process_model_facade_factory';
 export * from './process_token_facade';
